Updated the A10 PR tutorial design
Updated the A10 PR tutorial: it uses the new PR flow, which is easier that the scripted flow in 17.1. The flat designhasn't been changed. The unnecessary qsf and script files have been removed from the pr design. In addition, the "qsf" files have been changed based on the new PR flow.